Begin trying a buffer and atomic-xor RenderEngine + light blocking buffer of 32-bit ints ~ reorder light code to accomodate 32-light batching
shaders/shadowVolumeShader
+ pass through local light ID to fragment
+ write blocking data as bit in buffer with InterlockedXor


shaders/directionallight
+ pass through instance ID
+ check blocking data in buffer

+++ shaders/shadowVolumeShader_varying.def.hlsl
@@ -2,4 +2,4 @@ vec3 a_position  : POSITION;
 vec4 plane1 : TEXCOORD0 = vec4(0,0,0,0);
 vec4 plane2 : TEXCOORD1 = vec4(0,0,0,0);
 vec3 toLight : TEXCOORD5 = vec3(0,0,0);
-vec3 planeData : TEXCOORD6 = vec3(0,0,0);
+ivec3 planeData : TEXCOORD6 = ivec3(0,0,0);
